src/ecs/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
flask
requests
opentelemetry-exporter-otlp